+++ src/components/Description/DescriptionFull/components.tsx
@@ -5,6 +5,9 @@ import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import { materialDark } from "react-syntax-highlighter/dist/cjs/styles/prism";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { theme } from "@/app/config/themes";
+import CodeBlock from "@/components/CodeBlock";
+
+//criando componentes customizados para  parte de code 
 
   const components = {
 
@@ -43,75 +46,8 @@ import { theme } from "@/app/config/themes";
         {...props}
       />
     ),
-    code: ({ node, inline, className, children, ...props }: any) => {
-      const match = /language-(\w+)/.exec(className || "");
-      let codeString = String(children).replace(/\n$/, "");
-      ARRAY_SPECIAL_CHARS.forEach((item) => {
-        codeString = codeString.replace(
-          new RegExp(item.char, "g"),
-          item.replace
-        );
-      });
+    code: (props : any) => <CodeBlock {...props} />
 
-      codeString = String(codeString).replace(/\n$/, "");
-
-        function handleCopy(codeString: string): void {
-            throw new Error("Function not implemented.");
-        }
-
-      return !inline && match ? (
-        <Box
-          component="div"
-          sx={{
-            position: "relative",
-            cursor: "pointer",
-          }}
-          onClick={() => handleCopy(codeString)}
-        >
-          <SyntaxHighlighter
-            style={materialDark}
-            language={match[1]}
-            PreTag="div"
-            {...props}
-            showLineNumbers
-            customStyle={{
-              margin: 0,
-              borderRadius: "4px",
-            }}
-          >
-            {codeString}
-          </SyntaxHighlighter>
-          <Box
-            sx={{
-              position: "absolute",
-              top: "24px",
-              right: "16px",
-              cursor: "pointer",
-            }}
-            onClick={(e) => {
-              e.stopPropagation();
-              handleCopy(codeString);
-            }}
-          >
-            <ContentCopyIcon sx={{ color: "white" }} />
-          </Box>
-        </Box>
-      ) : (
-        <Typography
-          component="span"
-          sx={{
-            fontFamily: "monospace",
-            backgroundColor: "#f5f5f5",
-            padding: "2px 4px",
-            borderRadius: "4px",
-            wordBreak: "break-word",
-            whiteSpace: "pre-wrap",
-          }}
-        >
-          {children}
-        </Typography>
-      );
-    },
-  };
+    };
 
 export default components;
